vanilla/code-plagiarism-vanilla-02/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
transformers>=4.30.0
torch>=1.10.0
numpy>=1.21.0
faiss-cpu>=1.7.0
python-dotenv>=0.19.0
openai>=1.0.0
tqdm>=4.62.0
Jinja2>=3.0.0